[refactor] Move `LogStream` back to `Logger` now that most new APIs are in-place.
