+++ TODO.md
@@ -35,6 +35,7 @@
 
 ## Other Miscellaneous API changes
 
+- [x] Move `LogStream` back to `Logger`.
 - [x] Add LogStream.prototype.configure from `winston@2.0.0`
 - [x] `winston.Container` instances no longer add any transports by default.
