+++ lib/winston.js
@@ -35,8 +35,7 @@ winston.format    = require('./winston/formats');
 //
 winston.ExceptionHandler = require('./winston/exception-handler');
 winston.Container        = require('./winston/container').Container;
-winston.Logger           = require('./winston/log-stream');
-winston.LogStream        = require('./winston/log-stream');
+winston.Logger           = require('./winston/logger');
 winston.Transport        = require('winston-transport');
 
 //
@@ -61,7 +60,7 @@ winston.loggers = new winston.Container();
 //     winston.log('info', 'some message');
 //     winston.error('some error');
 //
-var defaultLogger = new winston.LogStream({
+var defaultLogger = new winston.Logger({
   //
   // TODO: Remove this transport and display a common
   // warning if logs are attempted to be written with
