+++ lib/winston/container.js
@@ -46,7 +46,7 @@ Container.prototype.get = Container.prototype.add = function (id, options) {
     //
     options.transports = existing ? existing.slice() : [];
 
-    this.loggers[id] = new winston.LogStream(options);
+    this.loggers[id] = new winston.Logger(options);
     this.loggers[id].on('close', function () {
       self._delete(id);
     });
