+++ lib/winston/logger.js
@@ -1,552 +1,528 @@
-/*
- * logger.js: Core logger object used by winston.
- *
- * (C) 2010 Charlie Robbins
- * MIT LICENCE
- *
- */
+'use strict';
 
-var events = require('events'),
+var stream = require('stream'),
     util = require('util'),
+    format = util.format,
     async = require('async'),
-    config = require('./config'),
-    common = require('./common'),
-    Stream = require('stream').Stream;
+    isStream = require('isstream'),
+    TransportStream = require('winston-transport'),
+    ExceptionHandler = require('./exception-handler'),
+    LegacyTransportStream = require('winston-transport/legacy'),
+    Profiler = require('./profiler'),
+    common = require('./common');
 
 const formatRegExp = /%[sdj%]/g;
 
-// //
-// // ### function Logger (options)
-// // #### @options {Object} Options for this instance.
-// // Constructor function for the Logger object responsible
-// // for persisting log messages and metadata to one or more transports.
-// //
-// var Logger = exports.Logger = function (options) {
-//   // events.EventEmitter.call(this);
-//   this.configure(options);
-// };
+/*
+ * Constructor function for the Logger object responsible
+ * for persisting log messages and metadata to one or more transports.
+ */
+var Logger = module.exports = function Logger(options) {
+  var self = this;
+  stream.Transform.call(this, { objectMode: true });
+  this.configure(options);
+};
 
 //
-// Inherit from `events.EventEmitter`.
+// Inherit from `stream.Transform`.
 //
-// util.inherits(Logger, events.EventEmitter);
+util.inherits(Logger, stream.Transform);
 
-// //
-// // ### function configure (options)
-// // This will wholesale reconfigure this instance by:
-// // 1. Resetting all transports. Older transports will be removed implicitly.
-// // 2. Set all other options including levels, colors, rewriters, filters,
-// //    exceptionHandlers, etc.
-// //
-// Logger.prototype.configure = function (options) {
-//   var self = this;
-//
-//   //
-//   // If we have already been setup with transports
-//   // then remove them before proceeding.
-//   //
-//   if (Array.isArray(this._names) && this._names.length) {
-//     this.clear();
-//   }
-//   //
-//   // Set Levels and default logging level
-//   //
-//   this.setLevels(options.levels);
-//
-//   //
-//   // Hoist other options onto this instance.
-//   //
-//   this.level       = options.level || 'info';
-//   this.exitOnError = typeof options.exitOnError !== 'undefined'
-//     ? options.exitOnError
-//     : true;
-//
-//   //
-//   // Setup internal state as empty Objects even though it is
-//   // defined lazily later to ensure a strong existential API contract.
-//   //
-//   this.exceptionHandlers = {};
-//   this.profilers         = {};
-//
-//   if (options.exceptionHandlers) {
-//     this.handleExceptions(options.exceptionHandlers);
-//   }
-// };
+/*
+ * ### function configure (options)
+ * This will wholesale reconfigure this instance by:
+ * 1. Resetting all transports. Older transports will be removed implicitly.
+ * 2. Set all other options including levels, colors, rewriters, filters,
+ *    exceptionHandlers, etc.
+ */
+Logger.prototype.configure = function (options) {
+  //
+  // Reset transports if we already have them
+  //
+  if (this.transports.length) {
+    this.clear();
+  }
 
-// //
-// // Inherit from `events.EventEmitter`.
-// //
-// util.inherits(Logger, events.EventEmitter);
+  //
+  // TODO: What happens if someone overwrites the format
+  // and then adds another TransportStream?
+  //
+  options = options || {};
+  this.format = options.format || this.format || require('./formats/default')();
 
-// //
-// // ### function log (level, msg, [meta], callback)
-// // #### @level {string} Level at which to log the message.
-// // #### @msg {string} Message to log
-// // #### @meta {Object} **Optional** Additional metadata to attach
-// // #### @callback {function} Continuation to respond to when complete.
-// // Core logging method exposed to Winston. Metadata is optional.
-// //
-// Logger.prototype.log = function (level) {
-//   var args = Array.prototype.slice.call(arguments, 1),
-//       self = this,
-//       transports;
-//
-//   while (args[args.length - 1] === null) {
-//     args.pop();
-//   }
-//
-//   //
-//   // Determining what is `meta` and what are arguments for string interpolation
-//   // turns out to be VERY tricky. e.g. in the cases like this:
-//   //
-//   //    logger.info('No interpolation symbols', 'ok', 'why', { meta: 'is-this' });
-//   //
-//   var callback  = typeof args[args.length - 1] === 'function'
-//     ? args.pop()
-//     : null;
-//
-//   //
-//   // Handle errors appropriately.
-//   //
-//   function onError(err) {
-//     if (callback) {
-//       callback(err);
-//     }
-//     else if (self.emitErrs) {
-//       self.emit('error', err);
-//     }
-//   }
-//
-//   if (this._names.length === 0) {
-//     return onError(new Error('Cannot log with no transports.'));
-//   }
-//   else if (typeof self.levels[level] === 'undefined') {
-//     return onError(new Error('Unknown log level: ' + level));
-//   }
-//
-//   //
-//   // If there are no transports that match the level
-//   // then be eager and return. This could potentially be calculated
-//   // during `setLevels` for more performance gains.
-//   //
-//   var targets = this._names.filter(function (name) {
-//     var transport = self.transports[name];
-//     return (transport.level && self.levels[transport.level] >= self.levels[level])
-//       || (!transport.level && self.levels[self.level] >= self.levels[level]);
-//   });
-//
-//   if (!targets.length) {
-//     if (callback) { callback(); }
-//     return;
-//   }
-//
-//   //
-//   // Determining what is `meta` and what are arguments for string interpolation
-//   // turns out to be VERY tricky. e.g. in the cases like this:
-//   //
-//   //    logger.info('No interpolation symbols', 'ok', 'why', { meta: 'is-this' });
-//   //
-//   var metaType  = Object.prototype.toString.call(args[args.length - 1]),
-//       fmtMatch  = args[0] && args[0].match && args[0].match(formatRegExp),
-//       isFormat  = fmtMatch && fmtMatch.length,
-//       validMeta = !isFormat
-//         ? metaType === '[object Object]' || metaType === '[object Error]' || metaType === '[object Array]'
-//         : metaType === '[object Object]',
-//       meta = validMeta ? args.pop() : {},
-//       msg  = util.format.apply(null, args);
-//
-//   //
-//   // Respond to the callback.
-//   //
-//   function finish(err) {
-//     if (callback) {
-//       if (err) return callback(err);
-//       callback(null, level, msg, meta);
-//     }
-//
-//     callback = null;
-//     if (!err) {
-//       self.emit('logged', level, msg, meta);
-//     }
-//   }
-//
-//   // If we should pad for levels, do so
-//   if (this.padLevels) {
-//     msg = new Array(this.levelLength - level.length + 1).join(' ') + msg;
-//   }
-//
-//   this.rewriters.forEach(function (rewriter) {
-//     meta = rewriter(level, msg, meta, self);
-//   });
-//
-//   this.filters.forEach(function(filter) {
-//     var filtered = filter(level, msg, meta, self);
-//     if (typeof filtered === 'string')
-//       msg = filtered;
-//     else {
-//       msg = filtered.msg;
-//       meta = filtered.meta;
-//     }
-//   });
-//
-//   //
-//   // For consideration of terminal 'color" programs like colors.js,
-//   // which can add ANSI escape color codes to strings, we destyle the
-//   // ANSI color escape codes when `this.stripColors` is set.
-//   //
-//   // see: http://en.wikipedia.org/wiki/ANSI_escape_code
-//   //
-//   if (this.stripColors) {
-//     var code = /\u001b\[(\d+(;\d+)*)?m/g;
-//     msg = ('' + msg).replace(code, '');
-//   }
-//
-//   //
-//   // Log for each transport and emit 'logging' event
-//   //
-//   function transportLog(name, next) {
-//     var transport = self.transports[name];
-//     transport.log(level, msg, meta, function (err) {
-//       if (err) {
-//         err.transport = transport;
-//         finish(err);
-//         return next();
-//       }
-//
-//       self.emit('logging', transport, level, msg, meta);
-//       next();
-//     });
-//   }
-//
-//   async.forEach(targets, transportLog, finish);
-//   return this;
-// };
+  //
+  // Hoist other options onto this instance.
+  //
+  this.setLevels(options.levels);
+  this.level = options.level || 'info';
+  this.exceptions = new ExceptionHandler(this);
+  this.profilers = {};
+  this.exitOnError = typeof options.exitOnError !== 'undefined'
+    ? options.exitOnError
+    : true;
 
-// //
-// // ### function query (options, callback)
-// // #### @options {Object} Query options for this instance.
-// // #### @callback {function} Continuation to respond to when complete.
-// // Queries the all transports for this instance with the specified `options`.
-// // This will aggregate each transport's results into one object containing
-// // a property per transport.
-// //
-// Logger.prototype.query = function (options, callback) {
-//   if (typeof options === 'function') {
-//     callback = options;
-//     options = {};
-//   }
-//
-//   var self = this,
-//       options = options || {},
-//       results = {},
-//       query = common.clone(options.query) || {},
-//       transports;
-//
-//   //
-//   // Helper function to query a single transport
-//   //
-//   function queryTransport(transport, next) {
-//     if (options.query) {
-//       options.query = transport.formatQuery(query);
-//     }
-//
-//     transport.query(options, function (err, results) {
-//       if (err) {
-//         return next(err);
-//       }
-//
-//       next(null, transport.formatResults(results, options.format));
-//     });
-//   }
-//
-//   //
-//   // Helper function to accumulate the results from
-//   // `queryTransport` into the `results`.
-//   //
-//   function addResults(transport, next) {
-//     queryTransport(transport, function (err, result) {
-//       //
-//       // queryTransport could potentially invoke the callback
-//       // multiple times since Transport code can be unpredictable.
-//       //
-//       if (next) {
-//         result = err || result;
-//         if (result) {
-//           results[transport.name] = result;
-//         }
-//
-//         next();
-//       }
-//
-//       next = null;
-//     });
-//   }
-//
-//   //
-//   // If an explicit transport is being queried then
-//   // respond with the results from only that transport
-//   //
-//   if (options.transport) {
-//     options.transport = options.transport.toLowerCase();
-//     return queryTransport(this.transports[options.transport], callback);
-//   }
-//
-//   //
-//   // Create a list of all transports for this instance.
-//   //
-//   transports = this._names.map(function (name) {
-//     return self.transports[name];
-//   }).filter(function (transport) {
-//     return !!transport.query;
-//   });
-//
-//   //
-//   // Iterate over the transports in parallel setting the
-//   // appropriate key in the `results`
-//   //
-//   async.forEach(transports, addResults, function () {
-//     callback(null, results);
-//   });
-// };
+  //
+  // Add all transports we have been provided.
+  //
+  if (options.transports) {
+    options.transports.forEach(function (transport) {
+      this.add(transport);
+    }, this);
+  }
 
-// //
-// // ### function stream (options)
-// // #### @options {Object} Stream options for this instance.
-// // Returns a log stream for all transports. Options object is optional.
-// //
-// Logger.prototype.stream = function (options) {
-//   var self = this,
-//       options = options || {},
-//       out = new Stream,
-//       streams = [],
-//       transports;
-//
-//   if (options.transport) {
-//     var transport = this.transports[options.transport];
-//     delete options.transport;
-//     if (transport && transport.stream) {
-//       return transport.stream(options);
-//     }
-//   }
-//
-//   out._streams = streams;
-//   out.destroy = function () {
-//     var i = streams.length;
-//     while (i--) streams[i].destroy();
-//   };
-//
-//   //
-//   // Create a list of all transports for this instance.
-//   //
-//   transports = this._names.map(function (name) {
-//     return self.transports[name];
-//   }).filter(function (transport) {
-//     return !!transport.stream;
-//   });
-//
-//   transports.forEach(function (transport) {
-//     var stream = transport.stream(options);
-//     if (!stream) return;
-//
-//     streams.push(stream);
-//
-//     stream.on('log', function (log) {
-//       log.transport = log.transport || [];
-//       log.transport.push(transport.name);
-//       out.emit('log', log);
-//     });
-//
-//     stream.on('error', function (err) {
-//       err.transport = err.transport || [];
-//       err.transport.push(transport.name);
-//       out.emit('error', err);
-//     });
-//   });
-//
-//   return out;
-// };
+  if (options.colors || options.emitErrs || options.formatters 
+    || options.padLevels || options.rewriters || options.stripColors) {
+    //
+    // TODO: Link to UPGRADE.md guide
+    //
+    throw new Error([
+      '{ colors, emitErrs, formatters, padLevels, rewriters, stripColors } were removed in winston@3.0.0.',
+      'Use a custom winston.format(function) instead.'
+    ].join('\n'));
+  }
 
-// //
-// // ### function handleExceptions ([tr0, tr1...] || tr0, tr1, ...)
-// // Handles `uncaughtException` events for the current process by
-// // ADDING any handlers passed in.
-// //
-// Logger.prototype.handleExceptions = function () {
-//   var args = Array.prototype.slice.call(arguments),
-//       handlers = [],
-//       self = this;
-//
-//   args.forEach(function (a) {
-//     if (Array.isArray(a)) {
-//       handlers = handlers.concat(a);
-//     }
-//     else {
-//       handlers.push(a);
-//     }
-//   });
-//
-//   this.exceptionHandlers = this.exceptionHandlers || {};
-//   handlers.forEach(function (handler) {
-//     self.exceptionHandlers[handler.name] = handler;
-//   });
-//
-//   this._hnames = Object.keys(self.exceptionHandlers);
-//
-//   if (!this.catchExceptions) {
-//     this.catchExceptions = this._uncaughtException.bind(this);
-//     process.on('uncaughtException', this.catchExceptions);
-//   }
-// };
+  if (options.exceptionHandlers) {
+    this.exceptions.handle(options.exceptionHandlers);
+  }
+};
 
-// //
-// // ### function unhandleExceptions ()
-// // Removes any handlers to `uncaughtException` events
-// // for the current process
-// //
-// Logger.prototype.unhandleExceptions = function () {
-//   var self = this;
-//
-//   if (this.catchExceptions) {
-//     Object.keys(this.exceptionHandlers).forEach(function (name) {
-//       var handler = self.exceptionHandlers[name];
-//       if (handler.close) {
-//         handler.close();
-//       }
-//     });
-//
-//     this.exceptionHandlers = {};
-//     Object.keys(this.transports).forEach(function (name) {
-//       var transport = self.transports[name];
-//       if (transport.handleExceptions) {
-//         transport.handleExceptions = false;
-//       }
-//     })
-//
-//     process.removeListener('uncaughtException', this.catchExceptions);
-//     this.catchExceptions = false;
-//   }
-// };
+/*
+ * @property {Array} Represents the current readableState
+ * pipe targets for this Logger instance.
+ */
+Object.defineProperty(Logger.prototype, 'transports', {
+  configurable: false,
+  enumerable: true,
+  get: function () {
+    var pipes = this._readableState.pipes;
+    return !Array.isArray(pipes)
+      ? [pipes].filter(Boolean)
+      : pipes;
+  }
+});
 
-// //
-// // ### function startTimer ()
-// // Returns an object corresponding to a specific timing. When done
-// // is called the timer will finish and log the duration. e.g.:
-// //
-// //    timer = winston.startTimer()
-// //    setTimeout(function(){
-// //      timer.done("Logging message");
-// //    }, 1000);
-// //
-// Logger.prototype.startTimer = function () {
-//   return new ProfileHandler(this);
-// };
+/*
+ * @property {Format} Current format associate with this instance.
+ * Handling this as a getter / setter allows us to implicitly
+ * remove the `readable` events on the format.
+ */
+Object.defineProperty(Logger.prototype, 'format', {
+  configurable: false,
+  enumerable: true,
+  get: function () {
+    return this._format;
+  },
+  set: function (fmt) {
+    if (this._format === fmt) {
+      //
+      // No-op on `stream.format = stream.format;`
+      //
+      return;
+    } else if (this._format) {
+      //
+      // If there has already been a format attached to this
+      // instance then remove our listener for `readable`
+      //
+      this._format.removeListener('readable', this._onReadableFormat);
+    }
+
+    //
+    // Listen to `readable` events on the format and then
+    // write those transformed `info` Objects onto
+    // the buffer for this instance.
+    //
+    // The original `info` Objects are written to the format
+    // in `Logger.prototype._transform`
+    //
+    var self = this;
+
+    //
+    // Keep a reference to our readable function
+    // so that we can remove the listener later.
+    //
+    this._onReadableFormat = function () {
+      var info;
+      while (null !== (info = self._format.read())) {
+        self.push(info);
+      }
+    };
+
+    this._format = fmt;
+    this._format.on('readable', this._onReadableFormat);
+  }
+});
+
+/*
+ * function log (level, msg, meta)
+ * function log (info)
+ * Ensure backwards compatibility with a `log` method
+ *
+ * Supports the existing API, which is now DEPRECATED:
+ *
+ *    logger.log('info', 'Hello world', { custom: true });
+ *
+ * And the new API with a single JSON literal:
+ *
+ *    logger.log({ level: 'info', message: 'Hello world', custom: true });
+ *
+ * @api deprecated
+ */
+Logger.prototype.log = function (level, msg, meta) {
+  //
+  // TODO: Check to see if arguments.length > 1 and
+  // if the string provided has any %% in it.
+  //
+  if (arguments.length === 1) {
+    return this.write(level);
+  }
+
+  meta = meta || {};
+  meta.level = level;
+  meta.message = msg;
+  this.write(meta);
+  return this;
+};
+
+/*
+ * @private function _transform (obj)
+ * Pushes data so that it can be picked up by all of
+ * our pipe targets.
+ */
+Logger.prototype._transform = function (info, enc, callback) {
+  //
+  // Remark: really not sure what to do here, but this has been
+  // reported as very confusing by pre winston@2.0.0 users as
+  // quite confusing when using custom levels.
+  //
+  if (!this.levels[info.level] && this.levels[info.level] !== 0) {
+    console.error('Unknown logger level: %s', info.level);
+  }
+
+  //
+  // Here we write to the `format` pipe-chain, which
+  // on `readable` above will push the formatted `info`
+  // Object onto the buffer for this instance.
+  //
+  // TODO: How do we handle TransportStream instances with their
+  // own format? We probably need two pipe chains here.
+  //
+  this._format.write(info);
+  callback();
+};
+
+/*
+ * function add (transport)
+ * Adds the transport to this logger instance by
+ * piping to it.
+ */
+Logger.prototype.add = function (transport) {
+  var self = this;
+
+  //
+  // Support backwards compatibility with all existing
+  // `winston@1.x.x` transport. All NEW transports should
+  // inherit from `winston.TransportStream`.
+  //
+  var target = !isStream(transport)
+    ? new LegacyTransportStream({ transport: transport })
+    : transport
+
+  if (!target._writableState || !target._writableState.objectMode) {
+    throw new Error('Transports must WritableStreams in objectMode. Set { objectMode: true }.');
+  }
+
+  //
+  // Listen for the `error` event on the new Transport
+  //
+  this._onError(target);
+  this.pipe(target);
+
+  if (transport.handleExceptions) {
+    this.exceptions.handle();
+  }
+
+  return this;
+};
+
+/*
+ * function remove (transport)
+ * Removes the transport from this logger instance by
+ * unpiping from it.
+ */
+Logger.prototype.remove = function (transport) {
+  var target = transport;
+  if (!isStream(transport)) {
+    target = this.transports.filter((match) => {
+      return match.transport === transport;
+    })[0];
+  }
+
+  if (target) { this.unpipe(target); }
+  return this;
+};
+
+/*
+ * function clear (transport)
+ * Removes all transports from this logger instance.
+ */
+Logger.prototype.clear = function () {
+  this.unpipe();
+  return this;
+};
+
+/*
+ * ### function close ()
+ * Cleans up resources (streams, event listeners) for all
+ * transports associated with this instance (if necessary).
+ */
+Logger.prototype.close = function () {
+  this.clear();
+  this.emit('close');
+  return this;
+};
+
+/*
+ * Sets the `target` levels specified on this instance.
+ * @param {Object} Target levels to use on this instance.
+ */
+Logger.prototype.setLevels = function (target) {
+  common.setLevels(this, this.levels, target);
+  return this;
+};
 
 //
-// ### function profile (id, [msg, meta, callback])
-// #### @id {string} Unique id of the profiler
-// #### @msg {string} **Optional** Message to log
-// #### @meta {Object} **Optional** Additional metadata to attach
-// #### @callback {function} **Optional** Continuation to respond to when complete.
+// ### function query (options, callback)
+// #### @options {Object} Query options for this instance.
+// #### @callback {function} Continuation to respond to when complete.
+// Queries the all transports for this instance with the specified `options`.
+// This will aggregate each transport's results into one object containing
+// a property per transport.
+//
+Logger.prototype.query = function (options, callback) {
+  if (typeof options === 'function') {
+    callback = options;
+    options = {};
+  }
+
+  var self = this,
+      options = options || {},
+      results = {},
+      query = common.clone(options.query) || {},
+      transports;
+
+  //
+  // Helper function to query a single transport
+  //
+  function queryTransport(transport, next) {
+    if (options.query) {
+      options.query = transport.formatQuery(query);
+    }
+
+    transport.query(options, function (err, results) {
+      if (err) {
+        return next(err);
+      }
+
+      next(null, transport.formatResults(results, options.format));
+    });
+  }
+
+  //
+  // Helper function to accumulate the results from
+  // `queryTransport` into the `results`.
+  //
+  function addResults(transport, next) {
+    queryTransport(transport, function (err, result) {
+      //
+      // queryTransport could potentially invoke the callback
+      // multiple times since Transport code can be unpredictable.
+      //
+      if (next) {
+        result = err || result;
+        if (result) {
+          results[transport.name] = result;
+        }
+
+        next();
+      }
+
+      next = null;
+    });
+  }
+
+  //
+  // Iterate over the transports in parallel setting the
+  // appropriate key in the `results`
+  //
+  async.forEach(this.transports.filter(function (transport) {
+    return !!transport.query;
+  }), addResults, function () {
+    callback(null, results);
+  });
+};
+
+//
+// ### function stream (options)
+// #### @options {Object} Stream options for this instance.
+// Returns a log stream for all transports. Options object is optional.
+//
+Logger.prototype.stream = function (options) {
+  options = options || {};
+
+  var self = this,
+      out = new stream.Stream(),
+      streams = [];
+
+  out._streams = streams;
+  out.destroy = function () {
+    var i = streams.length;
+    while (i--) streams[i].destroy();
+  };
+
+  //
+  // Create a list of all transports for this instance.
+  //
+  this.transports.filter(function (transport) {
+    return !!transport.stream;
+  }).forEach(function (transport) {
+    var stream = transport.stream(options);
+    if (!stream) return;
+
+    streams.push(stream);
+
+    stream.on('log', function (log) {
+      log.transport = log.transport || [];
+      log.transport.push(transport.name);
+      out.emit('log', log);
+    });
+
+    stream.on('error', function (err) {
+      err.transport = err.transport || [];
+      err.transport.push(transport.name);
+      out.emit('error', err);
+    });
+  });
+
+  return out;
+};
+
+//
+// ### function startTimer ()
+// Returns an object corresponding to a specific timing. When done
+// is called the timer will finish and log the duration. e.g.:
+//
+//    timer = winston.startTimer()
+//    setTimeout(function(){
+//      timer.done({ message: 'Logging message' });
+//    }, 1000);
+//
+Logger.prototype.startTimer = function () {
+  return new Profiler(this);
+};
+
+//
+// ### function profile (id, [info])
+// @param {string} id Unique id of the profiler
 // Tracks the time inbetween subsequent calls to this method
 // with the same `id` parameter. The second call to this method
 // will log the difference in milliseconds along with the message.
 //
-// Logger.prototype.profile = function (id) {
-//   var now = Date.now(), then, args,
-//       msg, meta, callback;
-//
-//   if (this.profilers[id]) {
-//     then = this.profilers[id];
-//     delete this.profilers[id];
-//
-//     // Support variable arguments: msg, meta, callback
-//     args     = Array.prototype.slice.call(arguments);
-//     callback = typeof args[args.length - 1] === 'function' ? args.pop() : null;
-//     meta     = typeof args[args.length - 1] === 'object' ? args.pop() : {};
-//     msg      = args.length === 2 ? args[1] : id;
-//
-//     // Set the duration property of the metadata
-//     meta.durationMs = now - then;
-//     return this.info(msg, meta, callback);
-//   }
-//   else {
-//     this.profilers[id] = now;
-//   }
-//
-//   return this;
-// };
+Logger.prototype.profile = function (id) {
+  var time = Date.now(),
+      timeEnd,
+      info,
+      args;
+
+  if (this.profilers[id]) {
+    timeEnd = this.profilers[id];
+    delete this.profilers[id];
+
+    //
+    // Attempt to be kind to users if they are still
+    // using older APIs.
+    //
+    args = Array.prototype.slice.call(arguments, 1);
+    if (typeof args[args.length - 1] === 'function') {
+      console.warn('Callback function no longer supported as of winston@3.0.0');
+      args.pop();
+    }
+
+    //
+    // Set the duration property of the metadata
+    //
+    info = typeof args[args.length - 1] === 'object' ? args.pop() : {};
+    info.level = info.level || 'info';
+    info.durationMs = time - timeEnd;
+    info.message = info.message || id;
+    return this.log(info);
+  }
+
+  this.profilers[id] = time;
+  return this;
+};
 
 //
-// ### function setLevels (target)
-// #### @target {Object} Target levels to use on this instance
-// Sets the `target` levels specified on this instance.
+// Some things stay the same
 //
-// Logger.prototype.setLevels = function (target) {
-//   return common.setLevels(this, this.levels, target);
-// };
+Logger.prototype.query = function () {
+  // More or less the same as winston@1.0.0
+};
+
+Logger.prototype.stream = function () {
+  // More or less the same as winston@1.0.0
+};
+
+/*
+ * Backwards compatibility to `exceptions.handle`
+ * in winston < 3.0.0.
+ *
+ * @api deprecated
+ */
+Logger.prototype.handleExceptions = function () {
+  console.warn('Deprecated: .handleExceptions() will be removed in winston@4. Use .exceptions.handle()');
+  var args = Array.prototype.slice.call(arguments);
+  this.exceptions.handle.apply(this.exceptions, args);
+};
+
+/*
+ * Backwards compatibility to `exceptions.handle`
+ * in winston < 3.0.0.
+ *
+ * @api deprecated
+ */
+Logger.prototype.unhandleExceptions = function () {
+  console.warn('Deprecated: .unhandleExceptions() will be removed in winston@4. Use .unexceptions.handle()');
+  var args = Array.prototype.slice.call(arguments);
+  this.exceptions.unhandle.apply(this.exceptions, args);
+};
+
+/*
+ * Throw a more meaningful deprecation notice
+ */
+Logger.prototype.cli = function () {
+  // TODO: Link to UPGRADE.md guide
+  throw new Error([
+    'Logger.cli() was removed in winston@3.0.0',
+    'Use a custom winston.format(function) instead.'
+  ].join('\n'));
+};
 
 //
-// ### @private function _uncaughtException (err)
-// #### @err {Error} Error to handle
-// Logs all relevant information around the `err` and
-// exits the current process.
+// ### @private function _onError (transport)
+// #### @transport {Object} Transport on which the error occured
+// #### @err {Error} Error that occurred on the transport
+// Bubbles the error, `err`, that occured on the specified `transport`
+// up from this instance if `emitErrs` has been set.
 //
-// Logger.prototype._uncaughtException = function (err) {
-//   var self = this,
-//       responded = false,
-//       info = exception.getAllInfo(err),
-//       handlers = this._getExceptionHandlers(),
-//       timeout,
-//       doExit;
-//
-//   //
-//   // Calculate if we should exit on this error
-//   //
-//   doExit = typeof this.exitOnError === 'function'
-//     ? this.exitOnError(err)
-//     : this.exitOnError;
-//
-//   function logAndWait(transport, next) {
-//     transport.logException('uncaughtException: ' + (err.message || err), info, next, err);
-//   }
-//
-//   function gracefulExit() {
-//     if (doExit && !responded) {
-//       //
-//       // Remark: Currently ignoring any exceptions from transports
-//       //         when catching uncaught exceptions.
-//       //
-//       clearTimeout(timeout);
-//       responded = true;
-//       process.exit(1);
-//     }
-//   }
-//
-//   if (!handlers || handlers.length === 0) {
-//     return gracefulExit();
-//   }
-//
-//   //
-//   // Log to all transports and allow the operation to take
-//   // only up to `3000ms`.
-//   //
-//   async.forEach(handlers, logAndWait, gracefulExit);
-//   if (doExit) {
-//     timeout = setTimeout(gracefulExit, 3000);
-//   }
-// };
+Logger.prototype._onError = function (transport) {
+  var self = this;
 
-//
-// ### @private function _getExceptionHandlers ()
-// Returns the list of transports and exceptionHandlers
-// for this instance.
-//
-// Logger.prototype._getExceptionHandlers = function () {
-//   var self = this;
-//
-//   return this._hnames.map(function (name) {
-//     return self.exceptionHandlers[name];
-//   }).concat(this._names.map(function (name) {
-//     return self.transports[name].handleExceptions && self.transports[name];
-//   })).filter(Boolean);
-// };
+  function transportError(err) {
+    self.emit('error', err, transport);
+  }
+
+  if (!transport.__winstonError) {
+    transport.__winstonError = transportError;
+    transport.on('error', transport.__winstonError);
+  }
+};
