+++ lib/winston/profiler.js
@@ -3,7 +3,7 @@
 //
 // ### @private Profiler
 // Constructor function for the Profiler instance used by
-// `LogStream.prototype.startTimer`. When done is called the timer
+// `Logger.prototype.startTimer`. When done is called the timer
 // will finish and log the duration.
 //
 var Profiler = module.exports = function Profiler(logger) {
