+++ test/fixtures/scripts/unhandle-exceptions.js
@@ -9,7 +9,7 @@
 var path = require('path'),
     winston = require('../../../lib/winston');
 
-var logger = new (winston.LogStream)({
+var logger = new (winston.Logger)({
   transports: [
     new (winston.transports.File)({
       filename: path.join(__dirname, '..', 'logs', 'unhandle-exception.log')
